Sliver detection and elimination:
After I clip a polygon against the more important stuff, I
check the area and minimum interior angle of each of it's contours.
If both (area and min interior angle) are below some threshhold I'm
calling it a sliver.  Then I go and look for another polygon such that
result = polygon_union( poly, sliver ) yields a result that doesn't
have anymore contours than the original poly.  This means the sliver
is adjacent to poly.  So I replace poly with result (the union) and
keep going.