Here is a wrap-up of the latest changes to JSBSim. It still is flaky, but much less so due to returning the aero reference point stuff to the config files. Don't know what happened there ...
Additionally, I have added a new field to the config file: CFG_VERSION. A
version number, currently 1.1, is assigned to the config file and a matching
definition is found in FGDefs.h. The two need to match. Tony has also added
code into FGAircraft.cpp to handle if aero reference point is not specified.