NAV1 is now the ILS-28R on airport KMYF, NAV2 is now the VOR radial 068 from MZB, ADF  is now the Compass locator on the outer marker. This combination is more than the legally required to fly any of KMYF-ILS-28R, KMYF-LOC-28R KMYF-NDB28. If you don't have access to the approach plates and would like them, let me know and I'll scan them (and put them on the webpage area).
The approaches do work; I've checked them all out in
terms of altitude profile, centerlines and other stuff.

In real life, the radar vectoring will basically abandon you
overhead KSEE airport at 4000 ft heading 210 or so.  Sometime
later you'll be turned to a heading of 260 if the controller
doesn't have too much else to do, just before you hit the
extended centerline.  You can't rely on that though.
Maintain 3500ft until established, 2100 ft until the outer marker,
If non-precision, maintain 1340 until crossing the radial,
then 900 thereafter until you miss, based on time from the NDB.
The miss takes you heading 270 to intercept a radial which this
hacky implementation will not let you set up the computer for.

The hacky math implementation does not take range and/or signal
strength into account, so you can fly to San Diego from England
by following the needle indication on the ADF.  It is also
fairly inaccurate math; about as accurate as the real-life signals.
When we have a _real_ radio module, I will be very happy to
throw all that code away.  For now, it makes it demonstratable.

Please notice the nastiness involving the "VARY_E" constant.
This is _not_ something that will go away with the radio module.
As far as I know, we don't have a routine that calculates
magnetic variation as a function of global position.
We will need one, probably within the next two months.