This is a somewhat rough first attempt at modelling cylinder head temperature.  The cylinder head is assumed to be at uniform temperature.  Obviously this is incorrect, but it simplifies things a lot, and we're just looking for the behaviour of CHT to be correct. Energy transfer to the cylinder head is assumed to be one third of the energy released by combustion at all conditions.  This is a reasonable estimate, although obviously in real life it varies with different conditions and possibly with CHT itself.  I've split energy transfer from the cylinder head into 2 terms - free convection - ie convection to stationary air, and forced convection, ie convection into flowing air.  The basic free convection equation is: dqdt = -hAdT   Since we don't know A and are going to set h quite arbitarily anyway I've knocked A out and just wrapped it up in h - the only real significance is that the units of h will be different but that dosn't really matter to us anyway.  In addition, we have the problem that the prop model I'm currently using dosn't model the backwash from the prop which will add to the velocity of the cooling air when the prop is turning, so I've added an extra term to try and cope with this.
In real life, forced convection equations are genarally empirically
derived, and are quite complicated and generally contain such things
as the Reynolds and Nusselt numbers to various powers.  The best
course of action would probably to find an empirical correlation from
the literature for a similar situation and try and get it to fit well.
 However, for now I am using my own made up very simple correlation
for the energy transfer from the cylinder head:

dqdt = -(h1.dT) -(h2.m_dot.dT) -(h3.rpm.dT)

where dT is the temperature different between the cylinder head and
the surrounding air, m_dot is the mass flow rate of cooling air
through an arbitary volume, rpm is the engine speed in rpm (this is
the backwash term), and h1, h2, h3 are co-efficients which we can play
with to attempt to get the CHT behaviour to match real life.

In order to change the values of CHT that the engine settles down at
at various conditions, have a play with h1, h2 and h3.  In order to
change the rate of heating/cooling without affecting equilibrium
values alter the cylinder head mass, which is really quite arbitary.
Bear in mind that altering h1, h2 and h3 will also alter the rate of
heating or cooling as well as equilibrium values, but altering the
cylinder head mass will only alter the rate.  It would I suppose be
better to read the values from file to avoid the necessity for
re-compilation every time I change them.