David Luff:  Here is an update to the engine model.  It now takes the actual air pressure and temperature from the LaRCSim model instead of assuming that it is at sea level as before.  This has reduced the ceiling from over 60000 ft to about 9000 ft.  This is a bit low (should be around 13 - 14000 ft I think) but I still have some stuff to do with the engine power correlation and its ignoring the temperature at the moment so I'm not panicking yet :-)
I've also changed the mixture-power correlation to one from a
published paper since the curve from the IO360 manual seemed to
be a load of rubbish, and didn't have any numbers on the mixture
axis anyway.

I've also knocked the full rich mixture down a touch in line with
Riley Rainey's recommendation, and cleaned up the code a bit.