Added --trace-read=<property> and --trace-write=<property> options. These will log trace messages whenever a specific property value is read or written through the property manager.  They do not cause FlightGear to poll bound values, so if a class variable that is bound is changed directly, no trace message will be displayed.
These options are also useful in conjunction with a debugger.
--trace-read will cause the private SimGear SGPropertyNode::trace_read
method to be invoked, and --trace-write will cause the private SimGear
SGPropertyNode::trace_write method to be invoked; in a debugger
session, a user can set breakpoints on these methods then get a
backtrace to see what specific points in the code are reading or
writing properties.