Attached are a fairly extensive series of patches to the ATC system.  A chap from Germany called Alexander Kappes (cc'd) got in touch with me a few weeks ago and has written the start of Approach control.  At the moment tuning in to a valid approach frequency (Dortmund or East Midlands) should result in vectors to a spot about 3 miles from the active runway, and a telling off if you stray too far from the correct course, in the console window.  He seems to know what he's doing so expect this to improve rapidly!!
I've added a rudimentry AI manager and a hardwired Cessna at
KEMT on the runway - I'll remove it before the next release if I don't
have it flying by then.  There seems to be an issue with framerate
which drops alarmingly when looking at it - I've a feeling that I've
possibly created several Cessnas on top of each other, but am not
sure.